Add files via upload
Purlsar Stars are a rare type of Neutron star that produce radio emission detectable here on Earth. They are of considerable scientific interest as probes space time , the interstella medium and states of matter , Machine learning tools are now being used to automatically label purlsar candidates to facilitate rapid analysis. Classification sysytem(SVM) in particular are being widely adopted, which treat the candidate datasets as binary classification problems.
    Attributes Information :
Each candidate is described by 8 continous variable, and a single class variable . The first four are simple statistics obtained from the integrated pulse profile(folded profile). This is an array of continous variable that describe a longitude-resolved version of the signal that has been averaged in both time and frequency. The remaining four variable are similarly obtained from the Dm-SNR curve.